move ray observability collection
